+++ tests/tests.rs
@@ -5,36 +5,22 @@
 // Copyright (c) DUSK NETWORK. All rights reserved.
 
 mod std_tests {
-    use dusk_bytes::ParseHexStr;
+    use dusk_bytes::{DeserializableSlice, ParseHexStr, Serializable};
     use dusk_pki::{PublicSpendKey, SecretSpendKey, ViewKey};
-    use rand::SeedableRng;
-    fn ssk_from_str(s: &str) -> SecretSpendKey {
-        use rand::rngs::StdRng;
-        use sha2::{Digest, Sha256};
-
-        let bytes = s.as_bytes();
-
-        let mut hasher = Sha256::default();
-        hasher.input(bytes);
-        let bytes = hasher.result();
-
-        let mut seed = [0u8; 32];
-        seed.copy_from_slice(&bytes[..32]);
-
-        SecretSpendKey::random(&mut StdRng::from_seed(seed))
-    }
+    use rand_core::OsRng;
 
     #[test]
     fn ssk_from_bytes() {
-        let ssk_a = ssk_from_str("some bytes");
-        let ssk_b = ssk_from_str("some bytes");
+        let ssk_a = SecretSpendKey::random(&mut OsRng);
+        let bytes = ssk_a.to_bytes();
+        let ssk_b = SecretSpendKey::from_slice(&bytes).expect("Serde error");
 
         assert_eq!(ssk_a, ssk_b);
     }
 
     #[test]
     fn keys_encoding() {
-        let ssk = ssk_from_str("some bytes");
+        let ssk = SecretSpendKey::random(&mut OsRng);
         let vk = ssk.view_key();
         let psk = ssk.public_spend_key();
 
@@ -53,15 +39,15 @@ mod std_tests {
     fn keys_consistency() {
         use dusk_jubjub::{JubJubScalar, GENERATOR_EXTENDED};
 
-        let r = JubJubScalar::random(&mut rand::thread_rng());
-        let ssk = SecretSpendKey::random(&mut rand::thread_rng());
+        let r = JubJubScalar::random(&mut OsRng);
+        let ssk = SecretSpendKey::random(&mut OsRng);
         let psk = ssk.public_spend_key();
         let vk = ssk.view_key();
         let sa = psk.gen_stealth_address(&r);
 
         assert!(vk.owns(&sa));
 
-        let wrong_ssk = SecretSpendKey::random(&mut rand::thread_rng());
+        let wrong_ssk = SecretSpendKey::random(&mut OsRng);
         let wrong_vk = wrong_ssk.view_key();
 
         assert_ne!(ssk, wrong_ssk);